Modificacion Primeras Pruebas

+++ SistemaTaxiMobil.Core/Entities/Parametros.cs
@@ -2,7 +2,7 @@
 {
     public class Parametros
     {
-        public byte ParametrosId { get; set; }
+        public int ParametrosId { get; set; }
         public decimal? TarifaBase { get; set; }
         public decimal? CostoMinuto { get; set; }
         public decimal? CostoKilometro { get; set; }
